Split clientaccess token/kubeconfig code
